requirements.txt
absl-py>=0.12.0
apache-beam>=2.36.0

